requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.5
django-classy-tags==4.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-taggit==3.1.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
Pillow==9.4.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
zipp==3.11.0
